refactor plans pages
